[inductor][triton] more JITCallable._hash_lock support
Follow-up to #161768.

It turns out that we do need to restore the _hash_lock, even in the non-triton_bundler case, for the MultiKernel case, where the hash lock is used.

To do this, we add `restore_after_unpickle()` which will restore fields (or if the old fields are not provided, initialize just the hash_lock)

ghstack-source-id: 799eb9ce4cfc4c064e08c0f132baa4522589aca4
Pull Request resolved: https://github.com/pytorch/pytorch/pull/162244