Used a fixed-length array to represent the interpreter stack
CPython represents the stack of a frame as a fixed length C array + a stack pointer. The pointer points to the next free slot. When the interpreter executes a PUSH or a POP operation, it increments or decrements the pointer, without having to resize the underlying array. This allows fast O(1) append/pop operations.

ghstack-source-id: ce0fa69d4f0340dd38aa60c01c7c521ff39d979e
Pull-Request: https://github.com/pytorch/pytorch/pull/162724